Extract components out from files page
